# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/exotic/address/{address}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves UTXOs which have exotic sat for a given address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx.exotic"
                ],
                "summary": "Retrieves UTXOs which have exotic sat for a given address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/wire.SatRangeUtxoResp"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/exotic/address/{address}/{type}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get UTXOs which is the specific exotic type in a address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx.exotic"
                ],
                "summary": "getExoticUtxosWithType",
                "parameters": [
                    {
                        "type": "string",
                        "description": "address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of SpecificExoticUtxo",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/wire.SpecificExoticUtxo"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/exotic/utxo/{utxo}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves all sat ranges and attributes in a given utxo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx.exotic"
                ],
                "summary": "Retrieves all sat ranges and attributes in a given utxo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "utxo",
                        "name": "utxo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/wire.SatRangeResp"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
...
        "/info/satributes": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the supported attributes of a sat",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx"
                ],
                "summary": "Retrieves the supported attributes of a sat",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/wire.SatributesResp"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
...
        "/sat/FindSatsInAddress/": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "find specific sats in address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx"
                ],
                "summary": "find specific sats in address",
                "parameters": [
                    {
                        "description": "address",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "sats",
                        "name": "sats",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "number"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/wire.SpecificSatResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
        "/sat/{sat}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves information about a sat based on the given sat ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx"
                ],
                "summary": "Retrieves information about a sat",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sat ID",
                        "name": "sat",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/wire.SatInfoResp"
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
...
        "/utxo/address/{address}/{value}": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get UTXOs in a address and its value is greater than the specific value. If value=0, get all UTXOs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordx"
                ],
                "summary": "Retrieves available UTXOs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "value",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/wire.PlainUtxosResp"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "wire.ExoticSatRangeUtxo": {
            "type": "object",
            "properties": {
                "sats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wire.SatDetailInfo"
                    }
                },
                "utxo": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
...
        "wire.PlainUtxo": {
            "type": "object",
            "properties": {
                "txid": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                },
                "vout": {
                    "type": "integer"
                }
            }
        },
        "wire.PlainUtxosResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wire.PlainUtxo"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "wire.SatDetailInfo": {
            "type": "object",
            "properties": {
                "block": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "satributes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "size": {
                    "type": "integer"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
        "wire.SatInfo": {
            "type": "object",
            "properties": {
                "cycle": {
                    "type": "integer"
                },
                "epoch": {
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "period": {
                    "type": "integer"
                },
                "sat": {
                    "type": "integer"
                },
                "satributes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "wire.SatInfoResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/wire.SatInfo"
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "wire.SatRange": {
            "type": "object",
            "properties": {
                "offset": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
        "wire.SatRangeResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/wire.ExoticSatRangeUtxo"
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "wire.SatRangeUtxoResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wire.ExoticSatRangeUtxo"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "wire.SatributesResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "wire.SpecificExoticUtxo": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "sats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wire.SatRange"
                    }
                },
                "type": {
                    "type": "string"
                },
                "utxo": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "wire.SpecificSat": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "sats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wire.SatRange"
                    }
                },
                "specificsat": {
                    "type": "integer"
                },
                "utxo": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "wire.SpecificSatResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wire.SpecificSat"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    }
}